Allow for a dynamic set of modules to be included
This will allow for various libraries to include itself into the
application controller.

Previously, I configured one such library to include itself in an
initializer. This didn't work quite well due to Rails' code reloader.
(The module would be included only in the version of the controller, and
since the initializer didn't re-run, the newly loaded controller did not
have the functionality provided by the module)

Note: I'm using a hash here so that it would be impossible for the same
module to be included twice. Ruby's hashes have the convenience of
conserving order, so it works quite nicely here.