Add "inline callbacks" for controllers
I learned just now that `after_action` runs after the view is rendered.
However, I needed to run some callbacks before the view was rendered,
but after the main body of the action. (Namely, I wanted to split out
pagination into callback)

This feature is enabled only on the `index` action for now; it should be
fairly easy to add this to the other actions, but there are some edge
cases that I need to think about before it's fully ready.

I skipped adding an explicit test, but it's partially implicitly tested
via the existing pagination tests.