Move "dashboard" inside its respective controller
Although I really like keeping the controller's configuration in a
separate class from the controller, I didn't really like how they're
split up in different files. When changing the admin page for a specific
resource, I felt that it's more natural to navigate to the controller.

Thus I merged the two files; the "dashboard" is now an inner class of
the controller. In addition, I renamed "dashboards" to "controls" to fit
with what I call them internally. (And in my mind, a dashboard is a page
that contains an overview of the state of your application.)

Since the `app/super` directory is no longer needed, the super installer
generator no longer creates it.

This is a breaking change.