Create Layout, Partial, and Action
The first two provide an interface for customizing views. The entrypoint
to both is the `#to_partial_path` method, which lets them be basically
interchangeable depending on what is necessary.

Action provides a way to customize controller actions and its respective
views. It supersedes Inline Callbacks, so I deleted it.

I also deleted View. It was a couple CSS utilities which was superseded
by `_super_layout.html.erb`.