Remove Super::Action framework
With the Super::Action framework, I was attempting to consolidate all
the configuration into one place so that one wouldn't need to (1) update
the Controls, (2) update the Controller, and (3) update the View.

I realized though that the Super::Action framework is weird. It isn't
intuitive, it isn't Rails. It's a step back.

So I'm removing it.