Make the index table headers "sticky" on the desktop
Sadly, `overflow-x-auto` (which allows for scrolling sideways if the
table is too wide) doesn't play nicely with `sticky`. Whenever
`overflow-x-auto` is applied, the `sticky` headers are automatically
disabled.

So on smaller screens, the table is horizontally scrollable, and on
larger screens, the table headers are sticky. I don't see any other
adverse effects.