Add flatpickr to forms
Flatpickr is a JavaScript library that provides a UI to set dates and
times. I chose it because it seemed relatively lightweight while having
the core features I needed.

One feature I wish it had is multi-timezone support. Flatpickr sends
timestamps in UTC while displaying the editing UI in the admin browser's
local timezone. This is very sensible, but doesn't make it too easy for
someone living in Los Angeles to set times for someone in New York.